refactored: search
