+++ srcs/ai/search/__init__.py
@@ -1,0 +1,1 @@
+from .search import search_memo
