+++ srcs/ai/search/_models/__init__.py
@@ -1,0 +1,1 @@
+from .memo import Memo
