+++ srcs/ai/search/utils/query_analyzer/__init__.py
@@ -1,0 +1,1 @@
+from .query_analyzer import query_analyzer
