+++ srcs/ai/search/utils/query_analyzer/chains/__init__.py
@@ -1,0 +1,1 @@
+from .query_analyzer_chain import query_analyzer_using_chain
