+++ srcs/ai/search/utils/query_analyzer/query_analyzer.py
@@ -1,6 +1,6 @@
 import logging
-from models.search import Search_query_type
-from ai.searching.utils.chains import query_analyzer_using_chain
+from routers._models.search import Search_query_type
+from ai.search.utils.query_analyzer.chains import query_analyzer_using_chain
 
 async def query_analyzer(query: str, lang: str="Korean") -> Search_query_type:
     analyzed_result=await query_analyzer_using_chain(query, lang)
