+++ srcs/ai/search/utils/regex/__init__.py
@@ -1,0 +1,1 @@
+from .regex import regex
