+++ srcs/ai/search/utils/regex/utils/__init__.py
@@ -1,0 +1,1 @@
+from .generate_regex import generate_regex
