+++ srcs/ai/search/utils/similarity/utils/agents/similarity_search_agent.py
@@ -1,10 +1,14 @@
-from typing import Any
+import json
+from typing import Any, Optional
 from langchain.agents import create_react_agent, AgentExecutor
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from ai.utils.llm import llm4o
 from langchain_core.prompts import PromptTemplate
-from ai.searching.utils.agents.similarity_search_agent_output_parser import Similarity_search_agent_output_parser
-from ai.searching.utils.agents.tools.internet_search import duckduckgo_search
+from typing import Union
+from langchain.agents.output_parsers import ReActJsonSingleInputOutputParser
+from langchain_core.output_parsers import PydanticOutputParser
+from langchain.schema import AgentAction, AgentFinish
+from ai.search.utils.similarity.utils.agents.tools import duckduckgo_search
 
 
 class _Similarity_search_agent_input(BaseModel):
@@ -12,10 +16,40 @@ class _Similarity_search_agent_input(BaseModel):
     lang: str
     question: str
 
+class Similarity_search_agent_output(BaseModel):
+    action: Optional[str]=None
+    action_input: Optional[dict[str, str]]=None
+    final_answer: Optional[str]=Field(description="answer to the user's question")
+    used_memo_ids: list[str]=Field(description="used memo ids", default=[])
+
 tools = [
     # memo_answerer,
     duckduckgo_search
 ]
+
+class Similarity_search_agent_output_parser(ReActJsonSingleInputOutputParser):
+    _pydantic_parser=PydanticOutputParser(pydantic_object=Similarity_search_agent_output)            
+        
+    def parse(self, llm_output: str) -> Union[AgentAction, AgentFinish]:
+        try:
+            parsed_output=self._pydantic_parser.parse(llm_output)
+            if parsed_output.action and parsed_output.action_input:
+                return AgentAction(parsed_output.action, parsed_output.action_input, llm_output)
+            else:
+                return AgentFinish(
+                    return_values={
+                        "answer": parsed_output.final_answer,
+                        "used_memo_ids": parsed_output.used_memo_ids
+                    },
+                    log=llm_output,
+                )
+        except json.JSONDecodeError:
+            raise ValueError(f"Invalid JSON output: {llm_output}")
+        except ValueError as e:
+            raise ValueError(f"Invalid output format: {str(e)}")
+
+    def get_format_instructions(self) -> str:
+        return self._pydantic_parser.get_format_instructions()
 
 _parser=Similarity_search_agent_output_parser()
