+++ srcs/ai/search/utils/similarity/utils/chains/similarity_result_with_memo_chain.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from operator import itemgetter
 from langchain_core.output_parsers import PydanticOutputParser
 from pydantic import BaseModel, Field
-from ai.searching._models.memo import Memo
+from ai.search._models.memo import Memo
 from ai.utils.llm import llm4o
 from langchain_core.prompts import PromptTemplate
