--- srcs/ai/searching/__init__.py
@@ -1,1 +1,0 @@
-from ai.searching.search import search_memo
