--- srcs/ai/searching/_models/__init__.py
@@ -1,1 +1,0 @@
-from ai.searching._models.memo import Memo
